+++ ui/app/src/components/ViewPackagesDialog/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2007-2022 Crafter Software Corporation. All Rights Reserved.
+ * Copyright (C) 2007-2025 Crafter Software Corporation. All Rights Reserved.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 3 as published by
@@ -14,8 +14,8 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export { default } from './WorkflowCancellationDialog';
+export { default } from './ViewPackagesDialog';
 
-export * from './WorkflowCancellationDialog';
+export * from './ViewPackagesDialog';
 
-export * from './WorkflowCancellationDialogUI';
+export * from './ViewPackagesDialogContainer';